% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinifyr.R
\name{tinify}
\alias{tinify}
\title{Compress your plot using Tinypng}
\usage{
tinify(file)
}
\arguments{
\item{file}{The filepath for the plot you want to compress (in PNG format)}
}
\description{
Running this function will compress your png using the Tinypng API (https://tinypng.com).
You will need to have a Tinypng API key set as an environment variable.
When run the function will save out your plot in its current location, with 'optimised-' in the file name.
}
\details{
To set your Tinypng API key as an environment variable write the following in R and then restart R to save it:
write_lines('TINYPNG_KEY = "your-api-key"', "~/.Renviron")

Keep in mind not to make this key public anywhere on Github and that you have just 500 free image compressions per month using the free Tinypng account.
}
\examples{
tinify(filepath)
}
\keyword{tinify}
